fix: fix installation reports completely for trials, update .env configuration for MikroTik host and enhance installation report view by removing obsolete columns and improving error messages

+++ internal/api/admin/asset/request.go
@@ -7,7 +7,7 @@ type CreateAdminAssetRequest struct {
 	SerialNumber string  `json:"serial_number" validate:"required"`
 	Date         string  `json:"date" validate:"required"`
 	CompanyID    *string `json:"company_id"`
-	Price        int64   `json:"price" validate:"required"`
+	Price        float64 `json:"price" validate:"required"`
 	Description  string  `json:"description"`
 	Site         string  `json:"site" validate:"required"`
 }
